Planet Lightsource hack
So instead for some normal lighting, I'll be adding planets and space shit in the scene, using the planet as a light source. With the character models, I can input them real soon. And the animation is gonna be fun especially the metal chrome liquid thing :DDDD
ANyWays yippie :D